+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/MetroAnnotations.kt
@@ -10,6 +10,7 @@ import dev.zacsweers.metro.compiler.ir.IrAnnotation
 import dev.zacsweers.metro.compiler.ir.IrMetroContext
 import dev.zacsweers.metro.compiler.ir.asIrAnnotation
 import dev.zacsweers.metro.compiler.ir.buildAnnotation
+import dev.zacsweers.metro.compiler.ir.findInjectableConstructor
 import dev.zacsweers.metro.compiler.ir.isAnnotatedWithAny
 import org.jetbrains.kotlin.fir.FirSession
 import org.jetbrains.kotlin.fir.expressions.FirAnnotationCall
@@ -25,6 +26,7 @@ import org.jetbrains.kotlin.fir.types.isResolved
 import org.jetbrains.kotlin.fir.types.resolvedType
 import org.jetbrains.kotlin.ir.declarations.IrAnnotationContainer
 import org.jetbrains.kotlin.ir.declarations.IrClass
+import org.jetbrains.kotlin.ir.declarations.IrConstructor
 import org.jetbrains.kotlin.ir.declarations.IrDeclaration
 import org.jetbrains.kotlin.ir.declarations.IrField
 import org.jetbrains.kotlin.ir.declarations.IrFunction
@@ -36,6 +38,7 @@ import org.jetbrains.kotlin.ir.symbols.IrSymbol
 import org.jetbrains.kotlin.ir.types.classOrNull
 import org.jetbrains.kotlin.ir.util.classId
 import org.jetbrains.kotlin.ir.util.deepCopyWithSymbols
+import org.jetbrains.kotlin.ir.util.parentAsClass
 
 @Poko
 internal class MetroAnnotations<T>(
@@ -301,37 +304,67 @@ private fun IrAnnotationContainer.metroAnnotations(
       yield(annotations)
 
       // You can fit so many annotations in properties
-      if (thisContainer is IrProperty) {
-        // Retrieve annotations from this property's various accessors
-        getter?.let { getter ->
-          if (getter != callingContainer) {
-            yield(getter.metroAnnotations(ids, callingContainer = thisContainer))
+      when (thisContainer) {
+        is IrProperty -> {
+          // Retrieve annotations from this property's various accessors
+          getter?.let { getter ->
+            if (getter != callingContainer) {
+              yield(getter.metroAnnotations(ids, callingContainer = thisContainer))
+            }
+          }
+          setter?.let { setter ->
+            if (setter != callingContainer) {
+              yield(setter.metroAnnotations(ids, callingContainer = thisContainer))
+            }
+          }
+          backingField?.let { field ->
+            if (field != callingContainer) {
+              yield(field.metroAnnotations(ids, callingContainer = thisContainer))
+            }
           }
         }
-        setter?.let { setter ->
-          if (setter != callingContainer) {
-            yield(setter.metroAnnotations(ids, callingContainer = thisContainer))
+
+        is IrSimpleFunction -> {
+          correspondingPropertySymbol?.owner?.let { property ->
+            if (property != callingContainer) {
+              val propertyAnnotations =
+                property.metroAnnotations(ids, callingContainer = thisContainer)
+              yield(propertyAnnotations)
+            }
           }
         }
-        backingField?.let { field ->
-          if (field != callingContainer) {
-            yield(field.metroAnnotations(ids, callingContainer = thisContainer))
+
+        is IrField -> {
+          correspondingPropertySymbol?.owner?.let { property ->
+            if (property != callingContainer) {
+              val propertyAnnotations =
+                property.metroAnnotations(ids, callingContainer = thisContainer)
+              yield(propertyAnnotations)
+            }
           }
         }
-      } else if (thisContainer is IrSimpleFunction) {
-        correspondingPropertySymbol?.owner?.let { property ->
-          if (property != callingContainer) {
-            val propertyAnnotations =
-              property.metroAnnotations(ids, callingContainer = thisContainer)
-            yield(propertyAnnotations)
+
+        is IrConstructor -> {
+          // Read from the class too
+          parentAsClass.let { parentClass ->
+            if (parentClass != callingContainer) {
+              val classAnnotations =
+                parentClass.metroAnnotations(ids, callingContainer = thisContainer)
+              yield(classAnnotations)
+            }
           }
         }
-      } else if (thisContainer is IrField) {
-        correspondingPropertySymbol?.owner?.let { property ->
-          if (property != callingContainer) {
-            val propertyAnnotations =
-              property.metroAnnotations(ids, callingContainer = thisContainer)
-            yield(propertyAnnotations)
+
+        is IrClass -> {
+          // Read from the inject constructor too
+          val constructor =
+            findInjectableConstructor(onlyUsePrimaryConstructor = false, ids.injectAnnotations)
+          if (constructor != null) {
+            if (constructor != callingContainer) {
+              val constructorAnnotations =
+                constructor.metroAnnotations(ids, callingContainer = thisContainer)
+              yield(constructorAnnotations)
+            }
           }
         }
       }