+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/InjectConstructorTransformer.kt
@@ -18,6 +18,7 @@ import dev.zacsweers.metro.compiler.ir.irExprBodySafe
 import dev.zacsweers.metro.compiler.ir.irInvoke
 import dev.zacsweers.metro.compiler.ir.irTemporary
 import dev.zacsweers.metro.compiler.ir.isExternalParent
+import dev.zacsweers.metro.compiler.ir.metroAnnotationsOf
 import dev.zacsweers.metro.compiler.ir.parameters.Parameter
 import dev.zacsweers.metro.compiler.ir.parameters.Parameters
 import dev.zacsweers.metro.compiler.ir.parameters.parameters
@@ -239,7 +240,11 @@ internal class InjectConstructorTransformer(
     // Generate a metadata-visible function that matches the signature of the target constructor
     // This is used in downstream compilations to read the constructor's signature
     val mirrorFunction =
-      generateMetadataVisibleMirrorFunction(factoryClass = factoryCls, target = targetConstructor)
+      generateMetadataVisibleMirrorFunction(
+        factoryClass = factoryCls,
+        target = targetConstructor,
+        metroAnnotationsOf(targetConstructor),
+      )
 
     factoryCls.dumpToMetroLog()